+++ packages/react-server/src/entry/browser.tsx
@@ -4,8 +4,10 @@ import type { RouterHistory } from "@tanstack/history";
 import React from "react";
 import ReactDOMClient from "react-dom/client";
 import { initializeReactClientBrowser } from "../features/client-component/browser";
-import { ErrorBoundary } from "../features/error/error-boundary";
-import { DefaultGlobalErrorPage } from "../features/error/global-error";
+import {
+  DefaultGlobalErrorPage,
+  ErrorBoundary,
+} from "../features/error/error-boundary";
 import {
   FlightDataContext,
   LayoutRoot,
